requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Django==5.0.4
django-crispy-forms==2.1
flatbuffers==24.3.25
fonttools==4.51.0
geographiclib==2.0
geopy==2.4.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
py==1.11.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
retry==0.9.2
retry-requests==2.0.0
scipy==1.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
